@@ -1,6 +1,7 @@
 {
   "name": "checkbox-group",
   "version": "1.0.0",
+  "license": "MIT",
   "description": "React Checkbox Group Component",
   "repository": "nackjicholson/checkbox-group",
   "author": {
@@ -16,59 +17,51 @@
   },
   "homepage": "https://github.com/nackjicholson/checkbox-group#readme",
   "devDependencies": {
-    "babel-cli": "^6.5.1",
-    "babel-core": "^6.5.1",
-    "babel-loader": "^6.2.2",
-    "babel-preset-es2015": "^6.5.0",
-    "babel-preset-react": "^6.5.0",
-    "babel-register": "^6.5.1",
-    "eslint": "^1.10.3",
-    "eslint-config-airbnb": "^5.0.0",
-    "eslint-plugin-react": "^3.16.1",
+    "babel-cli": "^6.14.0",
+    "babel-core": "^6.14.0",
+    "babel-loader": "^6.2.5",
+    "babel-preset-es2015": "^6.14.0",
+    "babel-preset-react": "^6.11.1",
+    "babel-register": "^6.14.0",
+    "eslint": "^3.4.0",
+    "eslint-config-xo-react": "^0.9.0",
+    "eslint-config-xo-space": "^0.14.0",
+    "eslint-plugin-react": "^6.1.2",
     "isparta": "^4.0.0",
-    "jscs": "^2.9.0",
-    "jsdom": "^8.0.2",
-    "mocha": "^2.4.5",
-    "react": "^0.14.7",
-    "react-dom": "^0.14.7",
-    "sinon": "^1.17.3",
-    "teaspoon": "^6.1.1",
-    "watch": "^0.17.1",
-    "webpack": "^1.12.13"
+    "jsdom": "^9.4.2",
+    "mocha": "^3.0.2",
+    "react": "^15.3.1",
+    "react-dom": "^15.3.1",
+    "sinon": "^1.17.5",
+    "teaspoon": "^6.4.1",
+    "watch": "^0.19.2",
+    "webpack": "^1.13.2"
   },
-  "dependencies": {
-    "lodash.includes": "^4.0.1"
-  },
+  "dependencies": {},
   "peerDependencies": {
-    "react": "^0.14.0",
-    "react-dom": "^0.14.0"
+    "react": "^15.3.1"
   },
   "scripts": {
-    "cov": "babel-node node_modules/.bin/isparta cover --report html --report text --excludes src/**/*.test.js _mocha -- --require src/test/jsdomHelper.js --reporter list src/**/*.test.js",
-    "lint": "eslint src && jscs src",
+    "cov": "babel-node node_modules/.bin/isparta cover --report html --report text --excludes src/**/*.test.js _mocha -- --require src/test/jsdomHelper.js src/**/*.test.js",
+    "lint": "eslint src",
     "build": "node build/buildLibrary.js",
     "test": "mocha --compilers js:babel-register --require src/test/jsdomHelper.js src/**/*.test.js",
     "tdd": "watch 'clear && npm run -s test' src",
     "tdd:lint": "watch 'clear && npm run -s lint && npm run -s test' src"
   },
   "eslintConfig": {
-    "extends": "airbnb",
-    "env": {
-      "mocha": true
-    },
+    "extends": [
+      "xo-space",
+      "xo-react/space"
+    ],
     "rules": {
-      "comma-dangle": 0
-    }
-  },
-  "jscsConfig": {
-    "preset": "airbnb",
-    "esnext": true,
-    "requireCurlyBraces": null,
-    "requireTrailingComma": null,
-    "validateQuoteMarks": {
-      "mark": "'",
-      "escape": true,
-      "ignoreJSX": true
+      "react/jsx-filename-extension": 0,
+      "react/require-optimization": 0
+    },
+    "env": {
+      "browser": true,
+      "node": true,
+      "mocha": true
     }
   },
   "babel": {
